Renamed static folder to assets.  Updated README.
Added first version of instructions for configuring server environment
for the server-side of SINQ.  These instructions are brief and will be
updated in the future.  For now, they require a basic knowledge of
Apache and Django and largley assume that both of these have been set
up on the server hardware.  Instructions include brief summary of
setting up the SINQ Django project and configuring Apache for the
project (i.e., so it is accessible from a web browser).